feat(requestvalidation): use getsync in validation
fix a race condition where deal status in request validation could fall behind actual events
dispatched
