+++ storagemarket/impl/provider.go
@@ -2,14 +2,11 @@ package storageimpl
 
 import (
 	"context"
-	"errors"
 	"io"
 
 	"github.com/hannahhoward/go-pubsub"
 	"github.com/ipfs/go-cid"
 	"github.com/ipfs/go-datastore"
-	"github.com/ipld/go-ipld-prime"
-	"github.com/libp2p/go-libp2p-core/peer"
 	"golang.org/x/xerrors"
 
 	"github.com/filecoin-project/go-address"
@@ -143,6 +140,11 @@ func NewProvider(net network.StorageMarketNetwork,
 
 	// register a data transfer event handler -- this will send events to the state machines based on DT events
 	dataTransfer.SubscribeToEvents(dtutils.ProviderDataTransferSubscriber(deals))
+
+	err = dataTransfer.RegisterVoucherType(&requestvalidation.StorageDataTransferVoucher{}, requestvalidation.NewUnifiedRequestValidator(&providerPushDeals{h}, nil))
+	if err != nil {
+		return nil, err
+	}
 
 	err = dataTransfer.RegisterTransportConfigurer(&requestvalidation.StorageDataTransferVoucher{}, dtutils.TransportConfigurer(&providerStoreGetter{h}))
 	if err != nil {
@@ -584,115 +586,6 @@ func providerDispatcher(evt pubsub.Event, fn pubsub.SubscriberFn) error {
 	}
 	cb(ie.evt, ie.deal)
 	return nil
-}
-
-// -------
-// providerDealEnvironment
-// -------
-
-type providerDealEnvironment struct {
-	p *Provider
-}
-
-func (p *providerDealEnvironment) Address() address.Address {
-	return p.p.actor
-}
-
-func (p *providerDealEnvironment) Node() storagemarket.StorageProviderNode {
-	return p.p.spn
-}
-
-func (p *providerDealEnvironment) Ask() storagemarket.StorageAsk {
-	sask := p.p.storedAsk.GetAsk()
-	if sask == nil {
-		return storagemarket.StorageAskUndefined
-	}
-	return *sask.Ask
-}
-
-func (p *providerDealEnvironment) DeleteStore(storeID multistore.StoreID) error {
-	return p.p.multiStore.Delete(storeID)
-}
-
-func (p *providerDealEnvironment) GeneratePieceCommitmentToFile(storeID *multistore.StoreID, payloadCid cid.Cid, selector ipld.Node) (cid.Cid, filestore.Path, filestore.Path, error) {
-	if p.p.universalRetrievalEnabled {
-		return providerutils.GeneratePieceCommitmentWithMetadata(p.p.fs, p.p.pio.GeneratePieceCommitmentToFile, p.p.proofType, payloadCid, selector, storeID)
-	}
-	pieceCid, piecePath, _, err := p.p.pio.GeneratePieceCommitmentToFile(p.p.proofType, payloadCid, selector, storeID)
-	return pieceCid, piecePath, filestore.Path(""), err
-}
-
-func (p *providerDealEnvironment) FileStore() filestore.FileStore {
-	return p.p.fs
-}
-
-func (p *providerDealEnvironment) PieceStore() piecestore.PieceStore {
-	return p.p.pieceStore
-}
-
-func (p *providerDealEnvironment) SendSignedResponse(ctx context.Context, resp *network.Response) error {
-	s, err := p.p.conns.DealStream(resp.Proposal)
-	if err != nil {
-		return xerrors.Errorf("couldn't send response: %w", err)
-	}
-
-	sig, err := p.p.sign(ctx, resp)
-	if err != nil {
-		return xerrors.Errorf("failed to sign response message: %w", err)
-	}
-
-	signedResponse := network.SignedResponse{
-		Response:  *resp,
-		Signature: sig,
-	}
-
-	err = s.WriteDealResponse(signedResponse)
-	if err != nil {
-		// Assume client disconnected
-		_ = p.p.conns.Disconnect(resp.Proposal)
-	}
-	return err
-}
-
-func (p *providerDealEnvironment) Disconnect(proposalCid cid.Cid) error {
-	return p.p.conns.Disconnect(proposalCid)
-}
-
-func (p *providerDealEnvironment) RunCustomDecisionLogic(ctx context.Context, deal storagemarket.MinerDeal) (bool, string, error) {
-	if p.p.customDealDeciderFunc == nil {
-		return true, "", nil
-	}
-	return p.p.customDealDeciderFunc(ctx, deal)
-}
-
-func (p *providerDealEnvironment) DealFunds() funds.DealFunds {
-	return p.p.dealFunds
-}
-
-func (p *providerDealEnvironment) TagPeer(id peer.ID, s string) {
-	p.p.net.TagPeer(id, s)
-}
-
-func (p *providerDealEnvironment) UntagPeer(id peer.ID, s string) {
-	p.p.net.UntagPeer(id, s)
-}
-
-var _ providerstates.ProviderDealEnvironment = &providerDealEnvironment{}
-
-type providerStoreGetter struct {
-	p *Provider
-}
-
-func (psg *providerStoreGetter) Get(proposalCid cid.Cid) (*multistore.Store, error) {
-	var deal storagemarket.MinerDeal
-	err := psg.p.deals.Get(proposalCid).Get(&deal)
-	if err != nil {
-		return nil, err
-	}
-	if deal.StoreID == nil {
-		return nil, errors.New("No store for this deal")
-	}
-	return psg.p.multiStore.Get(*deal.StoreID)
 }
 
 // ProviderFSMParameterSpec is a valid set of parameters for a provider FSM - used in doc generation
